IPOS

+++ application/models/Employee.php
@@ -40,7 +40,6 @@ class Employee extends Person
 
 		return $this->db->get();
 	}
-
 	/*
 	Gets information about a particular employee
 	*/
@@ -55,15 +54,41 @@ class Employee extends Person
 			return $query->row();
 		} else {
 			//Get empty base parent object, as $employee_id is NOT an employee
-			$person_obj = parent::get_info(-1);
+			//$person_obj = parent::get_info(-1);
 
 			//Get all the fields from employee table
 			//append those fields to base parent object, we we have a complete empty object
+			$person_obj = array();
 			foreach ($this->db->list_fields('employees') as $field) {
-				$person_obj->$field = '';
+				//$person_obj->$field = '';
+				$person_obj[$field] = "";
 			}
 
-			return $person_obj;
+			return (object) $person_obj;
+		}
+	}
+	public function get_info_by_authcode($auth_code)
+	{
+		$this->db->from('employees');
+		$this->db->join('people', 'people.person_id = employees.person_id');
+		$this->db->where('employees.auth_code', $auth_code);
+		$query = $this->db->get();
+
+		if ($query->num_rows() == 1) {
+			return $query->row();
+		} else {
+			//Get empty base parent object, as $employee_id is NOT an employee
+			//$person_obj = parent::get_info(-1);
+
+			//Get all the fields from employee table
+			//append those fields to base parent object, we we have a complete empty object
+			$person_obj = array();
+			foreach ($this->db->list_fields('employees') as $field) {
+				//$person_obj->$field = '';
+				$person_obj[$field] = "";
+			}
+
+			return (object) $person_obj;
 		}
 	}
 
@@ -131,6 +156,51 @@ class Employee extends Person
 
 		return $success;
 	}
+
+	/*
+	Checks if customer phone number exists
+	*/
+	public function check_phone_exists($phone, $customer_id = '')
+	{
+		// if the phone is empty return like it is not existing
+		if (empty($phone)) {
+			return FALSE;
+		}
+
+		$this->db->from('employees');
+		$this->db->join('people', 'people.person_id = employees.person_id');
+		$this->db->where('people.phone_number', $phone);
+		$this->db->where('employees.deleted', 0);
+
+		if (!empty($customer_id)) {
+			$this->db->where('employees.person_id !=', $customer_id);
+		}
+
+		return ($this->db->get()->num_rows() >= 1);
+	}
+
+	/*
+	Checks if customer email exists
+	*/
+	public function check_email_exists($email, $customer_id = '')
+	{
+		// if the email is empty return like it is not existing
+		if (empty($email)) {
+			return FALSE;
+		}
+
+		$this->db->from('employees');
+		$this->db->join('people', 'people.person_id = employees.person_id');
+		$this->db->where('people.email', $email);
+		$this->db->where('employees.deleted', 0);
+
+		if (!empty($customer_id)) {
+			$this->db->where('employees.person_id !=', $customer_id);
+		}
+
+		return ($this->db->get()->num_rows() >= 1);
+	}
+
 	public function saving_employee(&$person_data, &$grants_data, &$employee_data, $employee_id = FALSE)
 	{
 		$success = FALSE;
@@ -148,7 +218,8 @@ class Employee extends Person
 			}
 
 			//We have either inserted or updated a new employee, now lets set permissions. 
-			if ($success) {
+			//ignore if user is not new
+			if ($success && !$employee_id) {
 				//First lets clear out any grants the employee currently has.
 				$success = $this->db->delete('grants', array('person_id' => $employee_id));
 
@@ -170,7 +241,7 @@ class Employee extends Person
 
 		return $success;
 	}
-	public function save_employee_role(&$employee_data, &$grants_data, $employee_id = FALSE)
+	public function save_employee_role(&$employee_data, &$grants_data, $grants_module, $employee_id = FALSE)
 	{
 		$success = FALSE;
 
@@ -183,12 +254,21 @@ class Employee extends Person
 		//We have either inserted or updated a new employee, now lets set permissions. 
 		if ($success) {
 			//First lets clear out any grants the employee currently has.
+//            $already_granted = $this->db->get_where('grants',['person_id'=>$employee_id])->result();
+//            var_dump($already_granted);
+//            exit();
 			$success = $this->db->delete('grants', array('person_id' => $employee_id));
 
 			//Now insert the new grants
 			if ($success) {
+			    //created an index variable to match the grants data :::Lekan
+                //N.B it is expected that the grants data and grants module contain same number of elements, I try to make sure of that in the frontend
+                // Can't help it, it's the codebase I met and I can't just change it overnight so I work with it..:::Lekan
+                $cInd = 0;
 				foreach ($grants_data as $permission_id) {
-					$success = $this->db->insert('grants', array('permission_id' => $permission_id, 'person_id' => $employee_id));
+				    $grants_module_data = explode('_',$permission_id);
+					$success = $this->db->insert('grants', array('permission_id' => $permission_id, 'person_id' => $employee_id,'mod_id'=>$grants_module_data[0])); // added the table field (mod_id) :::Lekan
+                    $cInd++; // increment the index :::Lekan
 				}
 			}
 		}
@@ -293,6 +373,8 @@ class Employee extends Person
 		$this->db->group_start();
 		$this->db->like('first_name', $search);
 		$this->db->or_like('last_name', $search);
+		$this->db->or_like('username', $search);
+		$this->db->or_like('auth_code', $search);
 		$this->db->or_like('CONCAT(first_name, " ", last_name)', $search);
 		$this->db->group_end();
 		$this->db->where('deleted', 0);
@@ -349,6 +431,7 @@ class Employee extends Person
 		$this->db->or_like('email', $search);
 		$this->db->or_like('phone_number', $search);
 		$this->db->or_like('username', $search);
+		$this->db->or_like('auth_code', $search);
 		$this->db->or_like('CONCAT(first_name, " ", last_name)', $search);
 		$this->db->group_end();
 		//$this->db->where('deleted', 0);
@@ -372,6 +455,7 @@ class Employee extends Person
 		$this->db->or_like('email', $search);
 		$this->db->or_like('phone_number', $search);
 		$this->db->or_like('username', $search);
+		$this->db->or_like('auth_code', $search);
 		$this->db->or_like('CONCAT(first_name, " ", last_name)', $search);
 		$this->db->group_end();
 		//$this->db->where('deleted', 0);
@@ -478,11 +562,31 @@ class Employee extends Person
 		if ($permission_id == NULL) {
 			return TRUE;
 		}
-
 		$query = $this->db->get_where('grants', array('person_id' => $person_id, 'permission_id' => $permission_id), 1);
-
 		return ($query->num_rows() == 1);
 	}
+
+    /*
+     * function to get custom roles grants
+     * created by lekan :::Lekan
+     */
+    public function get_custom_grants($employee_id,$module){
+//        return [];
+        return $this->db->select("permission_id")
+            ->from('grants')
+//            ->like('permission_id',$module,'after')
+            ->where(['person_id'=>$employee_id,'mod_id'=>$module])
+//            ->group_by('mod_id')
+            ->get()->result_array();
+    }
+    public function can_vend($employee_id){
+        return $this->db->select("permission_id")
+            ->from('grants')
+//            ->like('permission_id',$module,'after')
+            ->where(['person_id'=>$employee_id,'permission_id'=>"irecharge"])
+//            ->group_by('mod_id')
+            ->get()->row();
+    }
 
 	/*
 	Gets employee permission grants
@@ -545,6 +649,24 @@ class Employee extends Person
 		}
 
 		return [];
+	}
+	public function get_employee_by_code_password($location, $code)
+	{
+
+		$this->db->select('employees.person_id,employees.username,employees.auth_code,employees.password,people.first_name,people.last_name');
+		$this->db->from('employees AS employees');
+		$this->db->join('people AS people', 'employees.person_id = people.person_id');
+		$this->db->where('branch_id', $location);
+		$this->db->where('auth_code', $code);
+		$this->db->where('deleted', 0);
+
+
+		$q = $this->db->get();
+		if ($q->num_rows() == 1) {
+			return $q->row();
+		}
+
+		return null;
 	}
 
 	public function get_employee_and_id()
